Do not delete kernel image from rootfs.
This append causes the kernel-image-uimage recipe to be empty. Some
images may want to keep the kernel image in boot, for example if they
want u-boot to load the kernel from the rootfs instead of a fat32 boot
partition. The deletion can be added as a postprocess command in the
image recipes that want it instead.