reintroduce xserver-xorg 1.19.6
Vivante xf86 driver crashes when loaded by current xserver 1.20.1
Waiting for the crash to be fixed, for var-som-mx6 we can still use
previous xserver working version 1.19.6
Also, this recipes is not required for not accelerated platforms,
so we can mask it for DART-6UL and VAR-SOM-MX7.

This patch must be removed once the xf86 driver crash will be fixed.